readded CONVERT_DEBEZIUM_DECIMAL_CONFIG as it can not be dropped without a major release

+++ kcbq-connector/src/main/java/com/wepay/kafka/connect/bigquery/convert/logicaltype/KafkaLogicalConverters.java
@@ -81,8 +81,7 @@ public class KafkaLogicalConverters {
     public DecimalConverter() {
       super(Decimal.LOGICAL_NAME,
           Schema.Type.BYTES,
-          LegacySQLTypeName.NUMERIC);
-
+          LegacySQLTypeName.FLOAT);
     }
 
     @Override
